JavaScript/jQuery rainbow text effect
All text between <a> tags on a page will, upon the page loading up, have
a "rainbow" shimmer effect that initiates when a user hovers over the
text. It was a good challenge and I learned a lot about how to access
the DOM elements and manipulate them with jQuery.